move named params into multi-optim
